requirements.txt
azure-devops==7.1.0b3
msrest==0.7.1
requests==2.31.0